+++ mobile/lib/features/wallet/send/send_lightning_screen.dart
@@ -1,6 +1,8 @@
 import 'dart:math';
 
+import 'package:bitcoin_icons/bitcoin_icons.dart';
 import 'package:flutter/material.dart';
+import 'package:get_10101/common/custom_app_bar.dart';
 import 'package:get_10101/common/amount_text.dart';
 import 'package:get_10101/common/application/channel_info_service.dart';
 import 'package:get_10101/common/application/switch.dart';
@@ -19,7 +21,6 @@ import 'package:get_10101/features/wallet/domain/wallet_type.dart';
 import 'package:get_10101/features/wallet/send/confirm_payment_modal.dart';
 import 'package:get_10101/features/wallet/wallet_change_notifier.dart';
 import 'package:get_10101/features/wallet/wallet_screen.dart';
-import 'package:go_router/go_router.dart';
 import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 
@@ -99,39 +100,7 @@ class _SendLightningScreenState extends State<SendLightningScreen> {
           child: Container(
             margin: const EdgeInsets.all(20.0),
             child: Column(crossAxisAlignment: CrossAxisAlignment.stretch, children: [
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  Expanded(
-                    child: Stack(
-                      children: [
-                        GestureDetector(
-                          child: Container(
-                              alignment: AlignmentDirectional.topStart,
-                              decoration: BoxDecoration(
-                                  color: Colors.transparent,
-                                  borderRadius: BorderRadius.circular(10)),
-                              width: 70,
-                              child: const Icon(
-                                Icons.arrow_back_ios_new_rounded,
-                                size: 22,
-                              )),
-                          onTap: () => GoRouter.of(context).pop(),
-                        ),
-                        const Row(
-                          mainAxisAlignment: MainAxisAlignment.center,
-                          children: [
-                            Text(
-                              "Send",
-                              style: TextStyle(fontWeight: FontWeight.w500, fontSize: 20),
-                            ),
-                          ],
-                        ),
-                      ],
-                    ),
-                  ),
-                ],
-              ),
+              const TenTenOneAppBar(title: "Send"),
               const SizedBox(
                 height: 20,
               ),
@@ -161,7 +130,7 @@ class _SendLightningScreenState extends State<SendLightningScreen> {
                       child: const Row(
                         mainAxisAlignment: MainAxisAlignment.spaceBetween,
                         children: [
-                          Icon(Icons.bolt, size: 14, color: Colors.white),
+                          Icon(BitcoinIcons.lightning, size: 14, color: Colors.white),
                           Text("Lightning", style: TextStyle(fontSize: 14, color: Colors.white))
                         ],
                       ),
@@ -243,7 +212,7 @@ class _SendLightningScreenState extends State<SendLightningScreen> {
                               const Row(
                                   mainAxisAlignment: MainAxisAlignment.spaceBetween,
                                   children: [
-                                    Icon(Icons.bolt, size: 18),
+                                    Icon(BitcoinIcons.lightning, size: 18),
                                     Text("Lightning", style: TextStyle(fontSize: 18))
                                   ]),
                               const SizedBox(height: 5),
@@ -253,7 +222,8 @@ class _SendLightningScreenState extends State<SendLightningScreen> {
                     ),
                     TenTenOneSwitch(
                         value: _payWithUsdp,
-                        showDisabled: false,
+                        isDisabled: usdpBalance <= 0,
+                        showDisabled: !_payWithUsdp,
                         onChanged: (value) => setState(() => _payWithUsdp = value)),
                     GestureDetector(
                       onTap: () => setState(() => _payWithUsdp = true),